Add more fined grained control over query selectors
Before we assumed everything was going to be textContent. But there are
some query selectors that dont need textcontent. Before the laconic
query selector used to look for the nth p tag. But then I saw that the
laconic is actually stored in the og:description tag of the page. So I
think that should be a lot more robust to draw from as the query
selector. But it does mean two things.

(1) We can't just use a query selector string because the og meta tag
laconic is stored in content so that "property" has to be generalised.