Add cached lookups and query selector property
This updates the way we fetch pages. Before we used to fetch every page
and rely on the browsers cache to store the request. But, chrome doesn't
actually store anything, for some reason, so now wwe explicitly cache
every request into a browser based cache. After that everything loads
pretty much instantly.

At some point we could cosnider just caching the trope object. But this
makes it harder to implement new features. What do we do if we want to
add images.

Now that we are using the new query selector + property definition, we
can easily grab the page image if it exists in the meta definition. But
we couldn't do that if we just "pickled" the Trope object. And the Dom
parsing is basically instant so I think this would just be premature
optimisation.

The other thing this function now does is following from the last
commit:

(2) We now have to check for the specific error message instead of just
truthy values. But still allow for null values to exist. The complex
return statement takes care of that.

See-Also: https://developer.mozilla.org/en-US/docs/Web/API/Cache