Bump tippy.js from 5.x to 6.x
This update includes breaking changes and
required adding the `appendTo: "parent"` config
option for the tippy instance, for some reason.
Without it the tippy popover would land in the
middle of the page for some reason. Probably
because this includes an update from the popper/core
library.

This also also updates out of date npm
dependencies.