Fixes #20
Seems like this is an issue with the chrome stylesheet was making
the tippy's root <div> display:block which caused text to break.

This probably wasn't a problem before because until we switched
to 6.x  the tippy box was not appended to the parent.

And I guess it dosn't happen in firefox because firefox doesn't have this
in its stylesheet.